# use bipf seekKeyCached()