--- a/test/basic.js
+++ b/test/basic.js
@@ -177,27 +177,34 @@ test('delete all', (t) => {
     db.publish(post2, (err) => {
       t.error(err, 'no err')
 
-      db.getJITDB().all(author(keys.id), 0, false, false, 'declared', (err, results) => {
-        t.error(err, 'no err')
-        t.equal(results.length, 30 + 5, 'got both new messages')
-
-        db.deleteFeed(keys.id, (err) => {
+      db.getJITDB().all(
+        author(keys.id),
+        0,
+        false,
+        false,
+        'declared',
+        (err, results) => {
           t.error(err, 'no err')
+          t.equal(results.length, 30 + 5, 'got both new messages')
 
-          db.getJITDB().all(
-            author(keys.id),
-            0,
-            false,
-            false,
-            'declared',
-            (err, results) => {
-              t.error(err, 'no err')
-              t.equal(results.length, 0, 'gone')
-              t.end()
-            }
-          )
-        })
-      })
+          db.deleteFeed(keys.id, (err) => {
+            t.error(err, 'no err')
+
+            db.getJITDB().all(
+              author(keys.id),
+              0,
+              false,
+              false,
+              'declared',
+              (err, results) => {
+                t.error(err, 'no err')
+                t.equal(results.length, 0, 'gone')
+                t.end()
+              }
+            )
+          })
+        }
+      )
     })
   })
 })