Context: https://github.com/ssb-ngi-pointer/jitdb/pull/208

Uses BIPF 1.6.0's `seekKeyCached` API instead of passing pValue around.

Pending on https://github.com/ssb-ngi-pointer/jitdb/pull/209 being merged first.