@arj03 2022-04-04
I ran the benchmarks and it seems as if caching is not kicking in for some reason. Havn't delved deep into the code to see why:

```
Only JITDB query:
master: 10.1s
this branch: 15.5s

Both JITDB + level indexes during query:
master: 37.4
this branch: 47.3
```