@arj03 2022-04-04
Well at least we have a good new baseline :) I can see that it is definitely hitting to cache inside bipf, so not exactly sure what is going on.