@staltz 2022-04-06
I ran benchmarks on my side as well, and can confirm that the new code would be slower. See before/after.

See also the heatmap from Chrome:

![Screenshot from 2022-04-06 12-53-49](https://user-images.githubusercontent.com/90512/161949670-24d2d497-82ee-475e-a90f-f294be37a2ea.png)

I think the root cause, if I can guess, is that there is too many `new Map()` going on, since the structure of the cache is WeakMap -> Map -> Map -> number. I don't know how to do this differently, because we can't just use `[buffer, start, target]` array as the key to the Map, because arrays aren't unique.

I figured that's it's worth documenting why this approach didn't work. But I'll close the PRs. :smiling_face_with_tear: 

## Before

```
query: 39678.767ms
```
```
query: 44708.553ms
```
```
query: 42647.140ms
```

## After

```
query: 47966.423ms
```
```
query: 47688.483ms
```
```
query: 48343.490ms
```