@arj03 2022-04-06
Yeah, good benchmarks. The multiple dictionaries would also have been my guess.